change migration type review_rate column to integer

+++ db/migrate/20210105141639_create_reviews.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 class CreateReviews < ActiveRecord::Migration[6.0]
   def change
     create_table :reviews do |t|
       t.belongs_to :user
       t.references :reviewable, polymorphic: true
-      t.float :review_rate
+      t.integer :review_rate
       t.text :review_text
       t.timestamps
     end
